BrokerMeta Copy method
