+++ kafkazk/brokers.go
@@ -11,27 +11,6 @@ const (
 	StubBrokerID int = int(^uint(0) >> 1)
 )
 
-// BrokerMetaMap is a map of broker IDs to BrokerMeta
-// metadata fetched from ZooKeeper. Currently, just
-// the rack field is retrieved.
-type BrokerMetaMap map[int]*BrokerMeta
-
-// BrokerMeta holds metadata that describes a broker,
-// used in satisfying constraints.
-type BrokerMeta struct {
-	StorageFree       float64 // In bytes.
-	MetricsIncomplete bool
-	// Metadata from ZooKeeper.
-	ListenerSecurityProtocolMap map[string]string `json:"listener_security_protocol_map"`
-	Endpoints                   []string          `json:"endpoints"`
-	Rack                        string            `json:"rack"`
-	JMXPort                     int               `json:"jmx_port"`
-	Host                        string            `json:"host"`
-	Timestamp                   string            `json:"timestamp"`
-	Port                        int               `json:"port"`
-	Version                     int               `json:"version"`
-}
-
 // BrokerMetricsMap holds a mapping of broker
 // ID to BrokerMetrics.
 type BrokerMetricsMap map[int]*BrokerMetrics
